superchain: contract implementations
Adds a concept of the contract implementations. yaml files that include
the semantic version to address exist for each network as well as a
global file for implementations that are deployed with create2. The
network specific config is merged into the global config, so that
contract addresses specific to a network will override the globals.
This will be the canonical list of contract addresses for implementations
going forward and will be used by the bundler tool to build upgrade
transactions.