Remove unnecessary information from frontier chains.
Based on the process at
https://github.com/ethereum-optimism/superchain-registry/blob/main/CONTRIBUTING.md#adding-a-frontier-chain,
we would not have created these files for these chains because they
are frontier chains, so removing them to reduce the maintainance
burden. We can always add them back if we add more requirements for
frontier chain information in the registry, or if they become a
standard chain.