Handle immutable vars when verifying L1 contract bytecodes (#497)
* Handle immutable vars when verifying L1 contract bytecodes

Values assigned to immutable vars at initialization are stored in bytecode. The coordinates of such stored immutable values are in the compiled contract AST/json (available in the monorepo). The same is now stored in a TOML config file as a stringified JSON for contracts that have immutables. The coordinates are used to mask the assigned value of the immutable vars in the bytecode, so that any hash that is computed on a deployed bytecode uses the same pre-image.

---------

Co-authored-by: Vinod Damle <vinod@oplabs.co>