ops: Add deployer package
Adds a `deployer` package to house code for upcoming changes to support multiple OP Deployer versions.

Now that OP Deployer only supports one contracts version per binary version, we can no longer import OP Deployer as a library directly into this package. Instead, we will need to shell out to different versions of the binary. This introduces a problem: the L2 Genesis component of the staging report needs to generate a standard intent, but must do so without importing the intent structs directly from the OP Deployer package.

To get around this, this PR imports `dasel` as a package and uses it to mutate intent and state files in-memory using Dasel's query syntax. This allows us to define what the standard intent and state files look like in native JSON/TOML and patch them with values from the user-supplied state without having to import the structs directly. The "standard" intent and state files the reports will use are defined in the `configs` directory, and are imported into the Go binary via `go:embed`.