Add V400-RC7 (#1055)
Adds a new tag, with addresses unchange from the previous.

This new release is necessary because of changes in the
StandardValidator, which although not listed in these toml files, is
deployed by op-deployer and thus needs a new tag in order to ensure the
correct version is deployed.

In the near future the SV will be added to the OPCM which will
implicitly include it here.